requirements.txt
langchain_groq
gradio
reportlab
python-dotenv
langchain